Refactor executive summary helpers into standalone module
